+++ src/main/java/org/sspoid/sspoid/api/controller/auth/KakaoLoginController.java
@@ -4,14 +4,12 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.sspoid.sspoid.api.dto.auth.KakaoLoginRequest;
 import org.sspoid.sspoid.api.dto.auth.KakaoUserInfoResponse;
-import org.sspoid.sspoid.api.service.KakaoLoginService;
+import org.sspoid.sspoid.api.service.auth.KakaoLoginService;
 
 @Slf4j
 @RestController
